Preserve parens around expressions in default export declaration that start with function or class (#6133)
