+++ tests/export_default/binary_and_template.js
@@ -1,0 +1,1 @@
+export default (function() {} + foo)``;
