+++ tests/export_default/class_instance.js
@@ -1,0 +1,1 @@
+export default (class {}.getInstance());
