+++ tests/export_default/function_in_template.js
@@ -1,0 +1,1 @@
+export default (function templ() {})`foo`;
