+++ tests/export_default/function_tostring.js
@@ -1,0 +1,1 @@
+export default (function() {}).toString();
