+++ tests/export_default/iife.js
@@ -1,0 +1,1 @@
+export default (function foo() {})();
