+++ tests/export_default/jsfmt.spec.js
@@ -1,1 +1,1 @@
-run_spec(__dirname, ["flow", "typescript"]);
+run_spec(__dirname, ["babel", "flow", "typescript"]);
