+++ tests/typescript_as/export_default_as.ts
@@ -1,0 +1,1 @@
+export default (function log() {} as typeof console.log)
